+++ SWMproject/SWMproject/SWMproject/Cards/Cards.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Identity;
+﻿using System.Collections.Generic;
 using Microsoft.Bot.Builder;
 using Microsoft.Bot.Schema;
