+++ SWMproject/SWMproject/SWMproject/Dialogs/OrderDialog.cs
@@ -4,7 +4,6 @@ using System.Threading.Tasks;
 using Microsoft.Bot.Builder;
 using Microsoft.Bot.Builder.Dialogs;
 using Microsoft.Bot.Builder.Dialogs.Choices;
-using Microsoft.Bot.Schema;
 using SWMproject.Data;
 
 namespace SWMproject.Dialogs
@@ -37,12 +36,12 @@ namespace SWMproject.Dialogs
                 AddiStepAsync,
                 //9 추가 요구사항
                 RequirementStepAsync,
-                //10 주문내역
-                SummaryStepAsync
+                ConfirmStepAsync
             };
 
             // Add named dialogs to the DialogSet. These names are saved in the dialog state.
             AddDialog(new AddToppingDialog(userState));
+            AddDialog(new OrderEndDialog(userState));
             AddDialog(new WaterfallDialog(nameof(WaterfallDialog), waterfallSteps));
             AddDialog(new TextPrompt(nameof(TextPrompt)));
             AddDialog(new ChoicePrompt(nameof(ChoicePrompt)));
@@ -133,8 +132,18 @@ namespace SWMproject.Dialogs
             return await stepContext.BeginDialogAsync(nameof(AddToppingDialog),null,cancellationToken);
         }
 
-        private static async Task<DialogTurnResult> WarmupStepAsync(WaterfallStepContext stepContext, CancellationToken cancellationToken)
+        private async Task<DialogTurnResult> WarmupStepAsync(WaterfallStepContext stepContext, CancellationToken cancellationToken)
         {
+            var orderData = await _orderDataAccessor.GetAsync(stepContext.Context, () => new OrderData(), cancellationToken);
+            if (stepContext.Result!= null)
+            {
+                if (orderData.Sandwiches.Count != 0)
+                {
+                    return await stepContext.ReplaceDialogAsync(nameof(OrderEndDialog), null, cancellationToken);
+                }
+                return await stepContext.EndDialogAsync();
+
+            }
             return await stepContext.PromptAsync(nameof(ChoicePrompt),
                new PromptOptions
                {
@@ -221,61 +230,13 @@ namespace SWMproject.Dialogs
             return await stepContext.PromptAsync(nameof(TextPrompt), new PromptOptions { Prompt = MessageFactory.Text("추가 요구사항을 입력하세요.") }, cancellationToken);
         }
 
-        private async Task<DialogTurnResult> SummaryStepAsync(WaterfallStepContext stepContext, CancellationToken cancellationToken)
+        private async Task<DialogTurnResult> ConfirmStepAsync(WaterfallStepContext stepContext, CancellationToken cancellationToken)
         {
-            stepContext.Values["requirement"] = (string)stepContext.Result;
-
             var orderData = await _orderDataAccessor.GetAsync(stepContext.Context, () => new OrderData(), cancellationToken);
-
+            stepContext.Values["requirement"] = stepContext.Result.ToString();
             orderData.Requirement = (string)stepContext.Values["requirement"];
-                
-            orderData.OrderNum++;
-            List<ReceiptItem> ItemList = new List<ReceiptItem> { new ReceiptItem(image:new CardImage(url: "https://www.subway.co.kr/images/common/logo_w.png")) };
-            ItemList.Add(new ReceiptItem("==========================="));
 
-            foreach (Sandwich tmpSand in orderData.Sandwiches)
-            {
-                ItemList.Add(new ReceiptItem("메뉴", price: Topping.menu_price[tmpSand.Menu].ToString() + "원"));
-                foreach (string temp in tmpSand.Topping)
-                {
-                    ItemList.Add(new ReceiptItem(temp, price: Topping.topping_price[temp].ToString() + "원"));
-                }
-                if (orderData.SetMenu != "단품")
-                {
-                    ItemList.Add(new ReceiptItem("세트", subtitle: $"{tmpSand.SetMenu},{tmpSand.SetDrink}", price: "1900원"));
-                    orderData.Price += 1900;
-                }
-                string vege = "";
-                string sauce = "";
-                foreach (string temp in tmpSand.Vege)
-                {
-                    vege += temp + ",";
-                }
-                foreach (string temp in tmpSand.Sauce)
-                {
-                    sauce += temp + ",";
-                }
-                ItemList.Add(new ReceiptItem("야채", subtitle: vege));
-                ItemList.Add(new ReceiptItem("소스", subtitle: sauce));
-                ItemList.Add(new ReceiptItem("-----------------------------------"));
-            }
-
-            ItemList.Add(new ReceiptItem("요구사항", subtitle: $"{orderData.Requirement}"));
-            ItemList.Add(new ReceiptItem("총 주문 개수", price: orderData.Num.ToString() + "개"));
-            ItemList.Add(new ReceiptItem("==========================="));
-
-            var receiptCard = new ReceiptCard
-            {
-                Title = "Subway receipt",
-                Facts = new List<Fact> { new Fact("주문번호", $"{orderData.OrderNum }"), new Fact("주문지점",$"{orderData.location}") }, //order number DB 연결?
-                Items = ItemList,
-                Total = $"{orderData.Price}원"
-            };
-            await stepContext.Context.SendActivityAsync(MessageFactory.Text("주문 내역을 확인해주세요."), cancellationToken);
-            await stepContext.Context.SendActivityAsync(MessageFactory.Attachment(receiptCard.ToAttachment()),cancellationToken);
-
-            orderData.Initial = true;
-            return await stepContext.EndDialogAsync(cancellationToken: cancellationToken);
+            return await stepContext.BeginDialogAsync(nameof(OrderEndDialog), null, cancellationToken);
         }
     }
 }
